+++ src/app/js/app-base.js
@@ -17,6 +17,9 @@ Provides a top-level application component which manages navigation and views.
 
 // TODO: Better handling of lifecycle for registered views:
 //
+//   * [!] Just redo basically everything with view management so there are no
+//     pre-`activeViewChange` side effects and handle the rest of these things:
+//
 //   * Seems like any view created via `createView` should listen for the view's
 //     `destroy` event and use that to remove it from the `_viewsInfoMap`. I
 //     should look at what ModelList does for Models as a reference.
@@ -25,6 +28,9 @@ Provides a top-level application component which manages navigation and views.
 //     needed if we have a `getView(name, create)` method, and already doing the
 //     above? We could do `app.getView('foo').destroy()` and it would be removed
 //     from the `_viewsInfoMap` as well.
+//
+//   * Should we wait to call a view's `render()` method inside of the
+//     `_attachView()` method?
 //
 //   * Should named views support a collection of instances instead of just one?
 //
@@ -301,6 +307,10 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
     Sets which view is active/visible for the application. This will set the
     app's `activeView` attribute to the specified `view`.
 
+    The `view` will be "attached" to this app, meaning it will be both rendered
+    into this app's `viewContainer` node and all of its events will bubble to
+    the app. The previous `activeView` will be "detached" from this app.
+
     When a string-name is provided for a view which has been registered on this
     app's `views` object, the referenced metadata will be used and the
     `activeView` will be set to either a preserved view instance, or a new
@@ -314,15 +324,17 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
     @example
         var app = new Y.App({
             views: {
-                users: {
+                usersView: {
                     // Imagine that `Y.UsersView` has been defined.
                     type: Y.UsersView
                 }
-            }
+            },
+
+            users: new Y.ModelList()
         });
 
         app.route('/users/', function () {
-            this.showView('users');
+            this.showView('usersView', {users: this.get('users')});
         });
 
         app.render();
@@ -330,26 +342,44 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
 
     @method showView
     @param {String|View} view The name of a view defined in the `views` object,
-        or a view instance.
+        or a view instance which should become this app's `activeView`.
     @param {Object} [config] Optional configuration to use when creating a new
-        view instance.
+        view instance. This config object can also be used to update an existing
+        or preserved view's attributes when `options.update` is `true`.
     @param {Object} [options] Optional object containing any of the following
         properties:
       @param {Function} [options.callback] Optional callback function to call
         after new `activeView` is ready to use, the function will be passed:
           @param {View} options.callback.view A reference to the new
             `activeView`.
-      @param {Boolean} [options.prepend=false] Whether the new view should be
+      @param {Boolean} [options.prepend=false] Whether the `view` should be
         prepended instead of appended to the `viewContainer`.
+      @param {Boolean} [options.render] Whether the `view` should be rendered.
+        **Note:** If no value is specified, a view instance will only be
+        rendered if it's newly created by this method.
+      @param {Boolean} [options.update=false] Whether an existing view should
+        have its attributes updated by passing the `config` object to its
+        `setAttrs()` method. **Note:** This option does not have an effect if
+        the `view` instance is created as a result of calling this method.
     @param {Function} [callback] Optional callback Function to call after the
         new `activeView` is ready to use. **Note:** this will override
-        `options.callback`. The function will be passed the following:
+        `options.callback` and it can be specified as either the third or fourth
+        argument. The function will be passed the following:
       @param {View} callback.view A reference to the new `activeView`.
     @chainable
     @since 3.5.0
     **/
     showView: function (view, config, options, callback) {
-        var viewInfo;
+        var viewInfo, created;
+
+        options || (options = {});
+
+        // Support the callback function being either the third or fourth arg.
+        if (callback) {
+            options.callback = callback;
+        } else if (Lang.isFunction(options)) {
+            options = {callback: options};
+        }
 
         if (Lang.isString(view)) {
             viewInfo = this.getViewInfo(view);
@@ -360,32 +390,38 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
             // `preserve` when detaching?
             if (viewInfo && viewInfo.preserve && viewInfo.instance) {
                 view = viewInfo.instance;
+
                 // Make sure there's a mapping back to the view metadata.
                 this._viewInfoMap[Y.stamp(view, true)] = viewInfo;
             } else {
-                view = this.createView(view, config);
-                view.render();
+                // TODO: Add the app as a bubble target during construction, but
+                // make sure to check that it isn't already in `bubbleTargets`!
+                // This will allow the app to be notified for about _all_ of the
+                // view's events. **Note:** This should _only_ happen if the
+                // view is created _after_ `activeViewChange`.
+
+                view    = this.createView(view, config);
+                created = true;
             }
         }
 
-        // TODO: Add `options.update` to update to view with the `config`, if
-        // needed. This could also call `setAttrs()` when the specified `view`
-        // already a View instance. Is this be too much overloading of the API?
-
-        // TODO: Add `options.render` to provide a way to control whether a view
-        // is rendered or not; by default, `render()` will only be called if
-        // this method created the View.
-
-        options || (options = {});
-
-        if (callback) {
-            options.callback = callback;
-        } else if (Lang.isFunction(options)) {
-            options = {callback: options};
+        // Update the specified or preserved `view` when signaled to do so.
+        // There's no need to updated a view if it was _just_ created.
+        if (options.update && !created) {
+            view.setAttrs(config);
         }
 
-        // TODO: Should the `callback` _always_ be called, even when the
-        // `activeView` does not change?
+        // TODO: Hold off on rendering the view until after it has been
+        // "attached", and move the call to render into `_attachView()`.
+
+        // When a value is specified for `options.render`, prefer it because it
+        // represents the developer's intent. When no value is specified, the
+        // `view` will only be rendered if it was just created.
+        if ('render' in options) {
+            options.render && view.render();
+        } else if (created) {
+            view.render();
+        }
 
         return this._set('activeView', view, {options: options});
     },
@@ -417,6 +453,9 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
 
         // TODO: Attach events here for persevered Views?
         // See related TODO in `_detachView`.
+
+        // TODO: Actually render the view here so that it gets "attached" before
+        // it gets rendered?
 
         // Insert view into the DOM.
         viewContainer[prepend ? 'prepend' : 'append'](view.get('container'));
@@ -644,7 +683,7 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
             // Force navigation to be enhanced and handled by the app when
             // `serverRouting` is falsy because the server might not be able to
             // properly handle the request.
-            Lang.isValue(options.force) || (options.force = true);
+            'force' in options || (options.force = true);
         }
 
         return PjaxBase.prototype._navigate.call(this, url, options);
@@ -723,8 +762,15 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
         after new `activeView` is ready to use, the function will be passed:
           @param {View} options.callback.view A reference to the new
             `activeView`.
-      @param {Boolean} [options.prepend=false] Whether the new view should be
+      @param {Boolean} [options.prepend=false] Whether the `view` should be
         prepended instead of appended to the `viewContainer`.
+      @param {Boolean} [options.render] Whether the `view` should be rendered.
+        **Note:** If no value is specified, a view instance will only be
+        rendered if it's newly created by this method.
+      @param {Boolean} [options.update=false] Whether an existing view should
+        have its attributes updated by passing the `config` object to its
+        `setAttrs()` method. **Note:** This option does not have an effect if
+        the `view` instance is created as a result of calling this method.
     @protected
     @since 3.5.0
     **/
@@ -819,7 +865,7 @@ App = Y.Base.create('app', Y.Base, [View, Router, PjaxBase], {
         @type View
         @default null
         @readOnly
-        @see AppBase.showView()
+        @see App.Base.showView()
         @since 3.5.0
         **/
         activeView: {