+++ src/app/HISTORY.md
@@ -4,6 +4,11 @@ App Framework Change History
 3.6.0
 -----
 
+### App
+
+* Added `render` and `update` options to the `showView()` method.
+  [PR #100 Pat Cavit]
+
 ### Router
 
 * The `req` object passed to routes now has a `pendingRoutes` property that